Bug 569048 authorize the server
Extract authentication expression evaluation code to be reused
