+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/internal/base/conditions/evaluation/BasePermissionEmittingService.java
@@ -16,25 +16,20 @@ import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Objects;
 
 import org.eclipse.passage.lic.internal.api.LicensedProduct;
 import org.eclipse.passage.lic.internal.api.LicensingException;
 import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
 import org.eclipse.passage.lic.internal.api.conditions.Condition;
 import org.eclipse.passage.lic.internal.api.conditions.ConditionPack;
-import org.eclipse.passage.lic.internal.api.conditions.EvaluationType;
 import org.eclipse.passage.lic.internal.api.conditions.ValidityPeriod;
 import org.eclipse.passage.lic.internal.api.conditions.ValidityPeriodClosed;
 import org.eclipse.passage.lic.internal.api.conditions.evaluation.Emission;
 import org.eclipse.passage.lic.internal.api.conditions.evaluation.ExpressionEvaluationException;
-import org.eclipse.passage.lic.internal.api.conditions.evaluation.ExpressionEvaluationService;
 import org.eclipse.passage.lic.internal.api.conditions.evaluation.ExpressionEvaluatorsRegistry;
 import org.eclipse.passage.lic.internal.api.conditions.evaluation.ExpressionParsingException;
 import org.eclipse.passage.lic.internal.api.conditions.evaluation.ExpressionPasringRegistry;
-import org.eclipse.passage.lic.internal.api.conditions.evaluation.ExpressionTokenAssessmentService;
 import org.eclipse.passage.lic.internal.api.conditions.evaluation.ExpressionTokenAssessorsRegistry;
-import org.eclipse.passage.lic.internal.api.conditions.evaluation.ParsedExpression;
 import org.eclipse.passage.lic.internal.api.conditions.evaluation.PermissionEmittingService;
 import org.eclipse.passage.lic.internal.api.diagnostic.Trouble;
 import org.eclipse.passage.lic.internal.api.diagnostic.TroubleCode;
@@ -50,20 +45,13 @@ import org.eclipse.passage.lic.internal.base.i18n.ConditionsEvaluationMessages;
 public final class BasePermissionEmittingService implements PermissionEmittingService {
 
 	private final StringServiceId id = new StringServiceId("default-emitter"); //$NON-NLS-1$
-	private final ExpressionPasringRegistry parsers;
-	private final ExpressionTokenAssessorsRegistry assessors;
-	private final ExpressionEvaluatorsRegistry evaluators;
+	private final Authentification authentification;
 
 	public BasePermissionEmittingService(//
 			ExpressionPasringRegistry parsers, //
 			ExpressionTokenAssessorsRegistry assessors, //
 			ExpressionEvaluatorsRegistry evaluators) {
-		Objects.requireNonNull(parsers);
-		Objects.requireNonNull(assessors);
-		Objects.requireNonNull(evaluators);
-		this.assessors = assessors;
-		this.parsers = parsers;
-		this.evaluators = evaluators;
+		authentification = new Authentification(parsers, assessors, evaluators);
 	}
 
 	@Override
@@ -99,25 +87,25 @@ public final class BasePermissionEmittingService implements PermissionEmittingSe
 	private ServiceInvocationResult<Emission> emitFor(Condition condition, ConditionPack pack,
 			LicensedProduct product) {
 		try {
-			expressionIsSatisfied(condition);
+			authentification.verify(condition.evaluationInstructions());
 		} catch (ExpressionParsingException e) {
 			return fail(pack, //
 					new LicenseInvalid(), //
-					String.format(ConditionsEvaluationMessages.getString("BasePermissionEmittingService.parse_failed"), // //$NON-NLS-1$
-							condition.evaluationInstructions().expression()), //
-					e);
-		} catch (ExpressionEvaluationException e) {
-			return fail(pack, //
-					new LicenseDoesNotMatch(), //
-					String.format(
-							ConditionsEvaluationMessages.getString("BasePermissionEmittingService.evaluation_failed"), // //$NON-NLS-1$
+					String.format(ConditionsEvaluationMessages.getString("BasePermissionEmittingService.parse_failed"), //$NON-NLS-1$
 							condition.evaluationInstructions().expression()), //
 					e);
 		} catch (LicensingException e) {
 			return fail(pack, //
 					new LicenseInvalid(), //
 					String.format(
-							ConditionsEvaluationMessages.getString("BasePermissionEmittingService.assessment_failed"), // //$NON-NLS-1$
+							ConditionsEvaluationMessages.getString("BasePermissionEmittingService.assessment_failed"), //$NON-NLS-1$
+							condition.evaluationInstructions().expression()), //
+					e);
+		} catch (ExpressionEvaluationException e) {
+			return fail(pack, //
+					new LicenseDoesNotMatch(), //
+					String.format(
+							ConditionsEvaluationMessages.getString("BasePermissionEmittingService.evaluation_failed"), //$NON-NLS-1$
 							condition.evaluationInstructions().expression()), //
 					e);
 		}
@@ -154,26 +142,6 @@ public final class BasePermissionEmittingService implements PermissionEmittingSe
 		return new BaseServiceInvocationResult<Emission>(//
 				new BaseDiagnostic(Collections.singletonList(new Trouble(code, explanation, e))), //
 				new Emission(pack));
-	}
-
-	private void expressionIsSatisfied(Condition condition)
-			throws ExpressionParsingException, ExpressionEvaluationException, LicensingException {
-		ExpressionTokenAssessmentService assessor = //
-				assessor(condition.evaluationInstructions().type());
-		ParsedExpression expression = new FormalizedExpression( //
-				condition.evaluationInstructions().expression(), //
-				parsers.get()).get();
-		ExpressionEvaluationService evaluator = evaluators.get().service(expression.protocol());
-		evaluator.evaluate(expression, assessor);
-	}
-
-	private ExpressionTokenAssessmentService assessor(EvaluationType type) throws LicensingException {
-		if (!assessors.get().hasService(type)) {
-			throw new LicensingException(String.format(
-					ConditionsEvaluationMessages.getString("BasePermissionEmittingService.no_assessment_service"), //$NON-NLS-1$
-					type));
-		}
-		return assessors.get().service(type);
 	}
 
 	private ZonedDateTime expiration(ValidityPeriod period) {
